requirements.txt
addict==2.4.0
appdirs==1.4.4
attrs==23.1.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
charset-normalizer==3.2.0
cityscapesScripts==2.2.2
click==8.1.6
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.2.7
cycler==0.11.0
Cython==3.0.0
exceptiongroup==1.1.2
filelock==3.12.2
flake8==5.0.4
fonttools==4.38.0
fsspec==2023.1.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==4.2.0
iniconfig==2.0.0
interrogate==1.5.0
isort==4.3.21
kiwisolver==1.4.4
markdown-it-py==2.2.0
matplotlib==3.5.3
mccabe==0.7.0
mdurl==0.1.2
mmcv-full==1.3.0
mmengine==0.8.4
mmsegmentation==1.1.1
numpy==1.21.6
opencv-python==4.8.0.76
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
platformdirs==3.10.0
pluggy==1.2.0
prettytable==3.7.0
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.16.1
pyparsing==3.1.1
pyquaternion==0.9.9
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rich==13.5.2
safetensors==0.3.2
scipy==1.7.3
six==1.16.0
tabulate==0.9.0
termcolor==2.3.0
terminaltables==3.1.10
timm==0.9.5
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
transformers==4.30.2
typing==3.7.4.3
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
xdoctest==1.1.1
yapf==0.33.0
zipp==3.15.0